requirements.txt
Flask
numpy
tensorflow
Pillow
keras